Move dummy midi note tracker closer to the tests (#169)
