--- src/midi/dummy_midi_note_tracker.rs
@@ -1,14 +1,0 @@
-use pm::types::Result;
-use midi::{TypedMidiMessage, MidiSink, MidiNoteTracker};
-
-pub struct DummyMidiNoteTracker;
-
-impl MidiNoteTracker for DummyMidiNoteTracker {
-    fn close_opened_notes(&mut self) {}
-}
-
-impl MidiSink for DummyMidiNoteTracker {
-    fn feed(&mut self, _: TypedMidiMessage) -> Result<()> {
-        Ok(())
-    }
-}
