+++ src/midi/mod.rs
@@ -10,14 +10,12 @@ use sdl2::rect::Rect;
 use measure::{Quant, QuantMidiEvent};
 
 mod port_midi_note_tracker;
-mod dummy_midi_note_tracker;
 mod midi_note_tracker;
 mod midi_sink;
 
 pub use self::port_midi_note_tracker::PortMidiNoteTracker;
 pub use self::midi_sink::MidiSink;
 pub use self::midi_note_tracker::MidiNoteTracker;
-pub use self::dummy_midi_note_tracker::DummyMidiNoteTracker;
 
 const NOTE_ON_STATUS: u8 = 0b10010000;
 const NOTE_OFF_STATUS: u8 = 0b10000000;
